+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
-name = "scwkms"
-version = "0.1.0"
+name = "scaleway_tink"
+version = "0.2.0"
 description = "Extension to Tink Python that provides Scaleway's Key Manager integration"
 authors = ["Scaleway <opensource@scaleway.com>"]
 readme = "README.md"
@@ -15,3 +15,9 @@ setuptools = "^70.0.0"
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[project.urls]
+Homepage = "https://github.com/scaleway/tink-py-scwkms"
+Documentation = "https://github.com/scaleway/tink-py-scwkms"
+Repository = "https://github.com/scaleway/tink-py-scwkms"
+Issues = "https://github.com/scaleway/tink-py-scwkms/issues"
